Fix the unresponsiveness in Unity in exiting program
This unresponsiveness exists when exiting and closing the TCP connection. Unity's MonoBehaviour methods, including Update and LateUpdate, run on the main thread. When exchangeTask.Wait() is called in the StopExchange function, it causes the main thread to wait for the exchangeTask to complete. This can result in freezing or unresponsiveness in Unity because the main thread is blocked.